pyproject.toml
[project]
name = "jenkins-mcp"
version = "0.1.0"
description = ""
authors = [
    {name = "tolatolatop",email = "tolatolatop@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastmcp>=2.0",
    "python-jenkins>=1.8.0",
]

[tool.poetry]
packages = [{include = "jenkins_mcp", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
